+++ database/multi.json
@@ -1,6 +1,5 @@
 {
-  "__comment__" : ". 양식은 아래를 따라야함.",
-
+  "__comment__": ". 양식은 아래를 따라야함.",
   "multi": [
     {
       "unique_key": "456XYZ_789QWE_2025-05-15T18:00:00Z",
@@ -24,5 +23,14 @@
     ],
     "ruleset": 1,
     "min_players": 3
+  },
+  "2025-06-23_05:25": {
+    "date": "2025-06-23 05:25",
+    "alert_time": "2025-06-23 04:55",
+    "members": [
+      "721290589340958761"
+    ],
+    "ruleset": 1,
+    "min_players": 2
   }
 }
